Update 2023-06-13-lee23a.md
I've noticed that Eunbyeol Cho, the principal author of the paper titled "Rediscovery of CNN’s Versatility for Text-based Encoding of Raw Electronic Health Records", is missing from the author list on your GitHub repository but is included in the PDF version. Please update the GitHub author list to correctly reflect this.